moved interfaces to data._base package, moved traj_cache and frames_from_file to util.
