+++ pyemma/coordinates/data/_base/datasource.py
