+++ pyemma/coordinates/data/_base/iterable.py
