+++ pyemma/coordinates/data/util/frames_from_file.py
