+++ pyemma/coordinates/data/util/traj_info_cache.py
